Add files via upload
This project is an application of the k-means clustering algorithm to the task of customer segmentation. Customer segmentation is the process of dividing a customer base into groups based on common characteristics, such as demographics, purchasing habits, and preferences. The k-means algorithm is a popular method for performing clustering, as it is simple and effective at finding clusters in the data.

The project begins by explaining the basics of the k-means algorithm and how it can be used for customer segmentation. It then presents a case study in which the k-means algorithm is applied to a dataset of customer data, including demographic information and purchasing habits.

Next, the project explores the results of the k-means clustering, including the identified clusters and their characteristics. The project also discusses the implications of the findings for marketing and customer relationship management.

The final section of the project presents the code and implementation details for the k-means algorithm, as well as a discussion of the limitations and potential improvements to the method.

Overall, this project is a valuable resource for anyone interested in learning more about the k-means algorithm and its application to customer segmentation.